requirements.txt
-e .

PyYAML
guessit
tmdbv3api